fixed overwriting columns and level discrepancy
We were updating the sale price of existing entries in the training data in the previous version, but have since updated the rbind to function properly.

Additionally, when using factor variables, the testing and training data levels did not match up, so we decided to use the levels that we had in the training data for consistency.